ability to convert some of the wander budget to scheduled urls
we might now about more urls for each domain than the ones that got a page centrality > 0.0. these urls won't be scheduled unless we convert some of the wander budget back to scheduled urls

+++ crates/core/src/webgraph/store_writer.rs
@@ -183,7 +183,7 @@ impl EdgeStoreWriter {
             .iter()
             .map(|p| IterableStoreReader::open(p).unwrap())
             .collect();
-        let file_reader = SortedIterableStoreReader::new(readers).map(|r| r.unwrap());
+        let file_reader = SortedIterableStoreReader::new(readers);
 
         let edges = std::mem::take(&mut self.edges);
 
@@ -212,7 +212,9 @@ impl EdgeStoreWriter {
 
 impl Drop for EdgeStoreWriter {
     fn drop(&mut self) {
-        std::fs::remove_dir_all(self.path.join("writer")).unwrap();
+        if self.path.join("writer").exists() {
+            std::fs::remove_dir_all(self.path.join("writer")).unwrap();
+        }
     }
 }
